Add file change services
Most of the COM interfaces of the file change service JTF.Run from a background thread and hence are not safe to use. The IVsAsyncFileChangeEx and IVsAsyncFileChangeEx2 are safe to use but they live in a namespace that is currently covered by this file and hence don't need to be opt'd out.